+++ src/main/java/com/aetherteam/enhanced_extinguishing/block/ExtinguishedLanternBlock.java
@@ -3,7 +3,7 @@ package com.aetherteam.enhanced_extinguishing.block;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.BlockGetter;
+import net.minecraft.world.level.LevelReader;
 import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.LanternBlock;
 import net.minecraft.world.level.block.state.BlockState;
@@ -20,7 +20,7 @@ public class ExtinguishedLanternBlock extends LanternBlock {
     }
 
     @Override
-    public ItemStack getCloneItemStack(BlockState state, HitResult target, BlockGetter level, BlockPos pos, Player player) {
+    public ItemStack getCloneItemStack(BlockState state, HitResult target, LevelReader level, BlockPos pos, Player player) {
         return Blocks.LANTERN.getCloneItemStack(state, target, level, pos, player);
     }
 }