+++ src/main/java/com/aetherteam/enhanced_extinguishing/block/ExtinguishedTorchBlock.java
@@ -28,7 +28,7 @@ public class ExtinguishedTorchBlock extends Block {
     }
 
     @Override
-    public ItemStack getCloneItemStack(BlockState state, HitResult target, BlockGetter level, BlockPos pos, Player player) {
+    public ItemStack getCloneItemStack(BlockState state, HitResult target, LevelReader level, BlockPos pos, Player player) {
         return Blocks.TORCH.getCloneItemStack(state, target, level, pos, player);
     }