+++ src/main/java/com/aetherteam/enhanced_extinguishing/data/providers/ExtinguishingRecipeProvider.java
@@ -1,9 +1,9 @@
 package com.aetherteam.enhanced_extinguishing.data.providers;
 
-import com.aetherteam.aether.recipe.AetherRecipeSerializers;
 import com.aetherteam.aether.recipe.builder.BiomeParameterRecipeBuilder;
+import com.aetherteam.aether.recipe.recipes.block.PlacementConversionRecipe;
 import com.aetherteam.nitrogen.recipe.BlockStateIngredient;
-import com.aetherteam.nitrogen.recipe.builder.BlockStateRecipeBuilder;
+import net.minecraft.core.HolderLookup;
 import net.minecraft.data.PackOutput;
 import net.minecraft.data.recipes.RecipeProvider;
 import net.minecraft.resources.ResourceLocation;
@@ -11,11 +11,13 @@ import net.minecraft.tags.TagKey;
 import net.minecraft.world.level.biome.Biome;
 import net.minecraft.world.level.block.Block;
 
+import java.util.concurrent.CompletableFuture;
+
 public abstract class ExtinguishingRecipeProvider extends RecipeProvider {
     private static String ID;
 
-    public ExtinguishingRecipeProvider(PackOutput output, String id) {
-        super(output);
+    public ExtinguishingRecipeProvider(PackOutput output, CompletableFuture<HolderLookup.Provider> lookupProvider, String id) {
+        super(output, lookupProvider);
         ID = id;
     }
 
@@ -23,7 +25,7 @@ public abstract class ExtinguishingRecipeProvider extends RecipeProvider {
         return new ResourceLocation(ID, name);
     }
 
-    protected static BlockStateRecipeBuilder convertPlacement(Block result, Block ingredient, TagKey<Biome> biome) {
-        return BiomeParameterRecipeBuilder.recipe(BlockStateIngredient.of(ingredient), result, biome, AetherRecipeSerializers.PLACEMENT_CONVERSION.get());
+    protected static BiomeParameterRecipeBuilder convertPlacement(Block result, Block ingredient, TagKey<Biome> biome) {
+        return BiomeParameterRecipeBuilder.recipe(BlockStateIngredient.of(ingredient), result, biome, PlacementConversionRecipe::new);
     }
 }